requirements.txt
streamlit
huggingface_hub
faiss-cpu
sentence-transformers
numpy